changed image uploading scheme

+++ src/pages/api/upload-image.js
@@ -1,13 +1,12 @@
-// pages/api/upload-image.js
 import { IncomingForm } from "formidable"
-import { PutObjectCommand } from "@aws-sdk/client-s3"
-import s3Client from "../../libs/aws-config"
 import fs from "fs"
 import path from "path"
+import { PutObjectCommand } from "@aws-sdk/client-s3"
+import s3Client from "../../libs/aws-config"
 
 export const config = {
   api: {
-    bodyParser: false, // Disables Next.js's default body parser to use Formidable
+    bodyParser: false, // Disable Next.js's default body parser to use Formidable
   },
 }
 
@@ -21,6 +20,7 @@ export default async function handler(req, res) {
     fs.mkdirSync(uploadDir, { recursive: true })
   }
 
+  // Set up the form with the specified upload directory
   const form = new IncomingForm({ uploadDir, keepExtensions: true })
 
   form.parse(req, async (err, fields, files) => {
@@ -30,33 +30,42 @@ export default async function handler(req, res) {
         .json({ message: "File upload error", error: err.toString() })
     }
 
+    // Access the uploaded image file
     const file = files.image && files.image[0]
+    if (!file) {
+      return res.status(400).json({ message: "Image file is required." })
+    }
 
     try {
-      const fileStream = fs.createReadStream(file.filepath)
-      const params = {
+      // Set S3 parameters for file upload
+      const s3Key = `uploads/${Date.now()}_${file.originalFilename}`
+      const uploadCommand = new PutObjectCommand({
         Bucket: process.env.S3_BUCKET_NAME,
-        Key: `uploads/${Date.now()}_${file.originalFilename}`,
-        Body: fileStream,
+        Key: s3Key,
+        Body: fs.createReadStream(file.filepath),
         ContentType: file.mimetype,
         ACL: "public-read",
-      }
+      })
 
-      await s3Client.send(new PutObjectCommand(params))
+      // Upload the file using the imported `s3Client`
+      await s3Client.send(uploadCommand)
 
-      // Delete the local file after uploading to S3
+      // Delete the local file after successfully uploading to S3
       fs.unlink(file.filepath, (unlinkErr) => {
         if (unlinkErr) {
           console.error("Error deleting local file:", unlinkErr)
         }
       })
 
-      const imageUrl = `https://${process.env.S3_BUCKET_NAME}.s3.${process.env.S3_REGION}.amazonaws.com/${params.Key}`
+      // Create the full image URL to send back in the response
+      const imageUrl = `https://${process.env.S3_BUCKET_NAME}.s3.${process.env.S3_REGION}.amazonaws.com/${s3Key}`
 
+      // Respond with the image URL and success message
       res
         .status(201)
         .json({ message: "Image uploaded successfully", url: imageUrl })
     } catch (error) {
+      // Log any errors that occur during the upload process
       console.error("Error during image upload:", error)
       res
         .status(500)
